ai-services/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
pymongo==4.10.1
pypdf==5.4.0
vaderSentiment==3.3.2
python-multipart==0.0.9
pandas==2.2.3
numpy
faiss-cpu==1.9.0
cohere==5.11.3
matplotlib==3.9.2
